requirements.txt
slackclient
apache-airflow[gcp_api,postgres]==1.10.6
google-api-python-client>=1.6.0,<2.0.0dev
google-auth-httplib2>=0.0.1
google-auth>=1.0.0,<2.0.0dev
google-cloud-bigtable==0.33.0
google-cloud-container>=0.1.1
google-cloud-dlp>=0.11.0
google-cloud-language>=1.1.1
google-cloud-spanner>=1.7.1,<1.10.0
google-cloud-storage~=1.16
google-cloud-translate>=1.3.3
google-cloud-videointelligence>=1.7.0
google-cloud-vision>=0.35.2
google-cloud-texttospeech>=0.4.0
google-cloud-speech>=0.36.3
grpcio-gcp>=0.2.2
httplib2~=0.9
pandas-gbq
PyOpenSSL
pytest